feat: Add device authorization grant (device code flow - rfc 8628) (#1539)
* Add Device model

This model represents the device session for the request and response stage
See section 3.1(https://datatracker.ietf.org/doc/html/rfc8628#section-3.1)
and 3.2(https://datatracker.ietf.org/doc/html/rfc8628#section-3.2)

* Adhere content-type request header to CGI standard

Django represents headers according to the common gateway interface(CGI)
standard. This means it's in all caps with words divided with a hyphen

However a lot of libraries follow the pattern of Something-Something
so this ensures the header is set correctly so libraries like oauthlib
can read it

* Add create device authorization response method

This method calls the server's create_device_authorization_response method
(https://datatracker.ietf.org/doc/html/rfc8628#section-3.2)

and is returns to the caller the information adhering to the rfc

* Update the grant type mapping to recognize device code

* Devices that are public should not need basic auth

The device flow is initiated by sending the client_id and and a scope.
This check should not fail if the client is public

* Add device settings

OAUTH_DEVICE_VERIFICATION_URI = the uri that comes back from the response
so the user knows where to go to. e.g example.com/device

OAUTH_DEVICE_USER_CODE_GENERATOR = Allows a custom callable to be passed in to control
how the user code is generated, stored in the db and returned back to the caller
DEVICE_MODEL = the device model

DEVICE_FLOW_INTERVAL = The time in seconds to wait before the device should poll again

* Create device authorization view

This view is to be used in an authorization server in order to provide
a /device endpoint

* Add the device user code form
* Add approve deny form
* Update token endpoint
* Use latest oauthlib 3.3.0