feat: add pluggable resource validator for resource servers
This validates the token audience (if there is one) against the request
URI.

Specs are unclear on *exactly* how this validation should be done.

Most implementations seem to require an *exact* match between the token
`aud` and the resource server identifier (Auth0, Okta, AWS Cognito)

This is a pain because it requires some extra configuration on the
resource server to define what exactly is 'the resource server identifier'
 - is it the host name, a hardcoded identifier or some subresource path?

I have opted for a pluggable system so the project can define what
approach to use, but I've chosen a default approach which I hope is
more flexible and requires no configuration - we match the token
audience claim to the request using a url prefix.

i.e. when requesting `https://example.com/users/foo`, a token with
`aud: ["https://example.com/users"]` would match.

This approach is implemented by Ory Hydra:
https://www.ory.com/docs/hydra/guides/audiences#audience-in-authorization-code-implicit-and-hybrid-flows

I also found a ticket requesting this feature for Ory Oathkeeper:
https://github.com/ory/oathkeeper/issues/656

Changes:
- Add `validate_resource_as_url_prefix()` with URL prefix matching logic
- New setting: `RESOURCE_SERVER_TOKEN_RESOURCE_VALIDATOR` (pluggable)
- Use the validator function in `validate_bearer_token()` when the token
  has an audience claim